[fix] 매칭 body 값 token 변경

+++ src/main/java/koscom/team6/domain/matching/dto/MatchingResponse.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class MatchingResponse {
 
-    int player1Id;
-    int player2Id;
+    Long player1Id;
+    Long player2Id;
     String matchSessionId;
 }
